+++ src/index.ts
@@ -1,7 +1,9 @@
-import "./headless-ui"
 import "./headless-popover"
 import "./headless-dropdown"
 import "./headless-navigation"
 import "./headless-disclosure"
-import "./headless-scrolltop"
+import "./headless-rating"
 import "./headless-darkmode"
+
+import "./headless-scrolltop"
+import "./headless-goback"