+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Release Notes
 
-## [Unreleased](https://github.com/designbycode/vanilla-headless/compare/v0.2.2...HEAD)
+## [Unreleased](https://github.com/designbycode/vanilla-headless/compare/v0.3.2...HEAD)
 - Unreleased
+
+## [v0.3.2](https://github.com/designbycode/vanilla-headless/compare/v0.3.1...v0.3.2) - 2022-10-15
+### Added
+- Added abstract HeadlessButton class to handle aria-pressed attribute
+- GoBack: added ``headless-goback`` directive that implements HeadlessButton
+### Changes
+- ScrollTop: implement HeadlessButton
 
 ## [v0.3.1](https://github.com/designbycode/vanilla-headless/compare/v0.2.2...v0.3.1) - 2022-10-14
 ### Added